+++ src/Workspace/Files/CpuPackageFile.ts
@@ -1,10 +1,10 @@
-import { Element as XmlElement } from '@oozcitak/dom/lib/dom/interfaces';
+import * as vscode from 'vscode';
 import { Uri } from 'vscode';
 import { splitShellArgs } from '../../Tools/Helpers';
 import { logger } from '../../Tools/Logger';
 import { pathResolve, winPathToPosixPath } from '../../Tools/UriTools';
-import { getChildElements } from '../../Tools/XmlDom';
 import { AsPackageFile } from './AsPackageFile';
+import { ParsedXmlObject } from './AsXmlParser';
 
 /** CPU and build configuration data */
 export interface CpuConfiguration {
@@ -40,52 +40,40 @@ export class CpuPackageFile extends AsPackageFile {
      * @param filePath The Cpu package file path. e.g. `C:\Projects\Test\Physical\TestConfig\TestPLC\Cpu.pkg`
      * @returns The Cpu package file representation which was parsed from the file
      */
-    public static async createFromPath(filePath: Uri): Promise<CpuPackageFile | undefined> {
+    public static async createFromFile(filePath: Uri): Promise<CpuPackageFile | undefined> {
         // Create and initialize object
         try {
-            const xmlFile = new CpuPackageFile(filePath);
-            await xmlFile._initialize();
-            return xmlFile;
+            const textDoc = await vscode.workspace.openTextDocument(filePath);
+            const fileContent = textDoc.getText();
+            return new CpuPackageFile(filePath, fileContent);
         } catch (error) {
             if (error instanceof Error) {
-                logger.error(`Failed to read Cpu package file from path '${filePath.fsPath}': ${error.message}`); //TODO uri log #33
+                logger.error(`Failed to read Cpu package file from path "${filePath.fsPath}": ${error.message}`); //TODO uri log #33
             } else {
-                logger.error(`Failed to read Cpu package file from path '${filePath.fsPath}'`); //TODO uri log #33
+                logger.error(`Failed to read Cpu package file from path "${filePath.fsPath}"`); //TODO uri log #33
             }
             logger.debug('Error details:', { error });
             return undefined;
         }
     }
 
-    /** Object is not ready to use after constructor due to async operations,
-     * _initialize() has to be called for the object to be ready to use! */
-    protected constructor(filePath: Uri) {
-        super(filePath);
+    /** TODO doc */
+    protected constructor(filePath: Uri, fileContent: string) {
+        super(filePath, fileContent);
         // other properties rely on async and will be initialized in #initialize()
-    }
-
-    /**
-     * Async operations to finalize object construction
-     * @throws If a required initialization process failed
-     */
-    protected async _initialize(): Promise<void> {
-        await super._initialize();
         if (this.type !== 'Cpu') {
             throw new Error('Root element name is not <Cpu>');
         }
-        this.#cpuConfig = getCpuConfiguration(this.rootElement);
+        this.#cpuConfig = getCpuConfiguration(this.xmlRootObj);
         // init done
-        this.#isInitialized = true;
-        this.#logWarningsOnInitialized();
+        this.#logWarningsForMissingProperties();
     }
-    #isInitialized = false;
 
     /** CPU and build configuration data */
     public get cpuConfig(): CpuConfiguration {
-        if (!this.#isInitialized || !this.#cpuConfig) { throw new Error(`Use of not initialized ${CpuPackageFile.name} object`); }
         return this.#cpuConfig;
     }
-    #cpuConfig: CpuConfiguration | undefined;
+    #cpuConfig: CpuConfiguration;
 
     /** toJSON required as getter properties are not shown in JSON.stringify() otherwise */
     public toJSON(): any {
@@ -95,7 +83,7 @@ export class CpuPackageFile extends AsPackageFile {
     }
 
     /** Check object properties and write warnings to the logger if important properties are missing */
-    #logWarningsOnInitialized() {
+    #logWarningsForMissingProperties() {
         // Collect missing properties
         const missingProperties: string[] = [];
         if (this.cpuConfig.cpuModuleId === undefined || this.cpuConfig.cpuModuleId.length === 0) {
@@ -110,7 +98,7 @@ export class CpuPackageFile extends AsPackageFile {
         // warn if there are missing properties
         if (missingProperties.length > 0) {
             const missingString = missingProperties.join(', ');
-            logger.warning(`Cpu package file '${this.filePath.toString(true)} does not contain values: ${missingString}`); //TODO uri log #33
+            logger.warning(`Cpu package file "${this.filePath.toString(true)}" does not contain values: ${missingString}`); //TODO uri log #33
         }
     }
 }
@@ -119,37 +107,36 @@ export class CpuPackageFile extends AsPackageFile {
  * Get all child objects from the package XML
  * @throws If there are multiple child object root nodes (<Files> or <Objects>)
  */
-function getCpuConfiguration(rootElement: XmlElement): CpuConfiguration {
+function getCpuConfiguration(rootElement: ParsedXmlObject): CpuConfiguration {
+    const rootAny = rootElement as any;
     // Get properties from <Configuration> element
-    const configElement = getChildElements(rootElement, 'Configuration').pop();
-    if (!configElement) {
+    const configElement = rootAny.Configuration;
+    if (typeof configElement !== 'object' || configElement === null) {
         throw new Error('No <Configuration> element found');
     }
-    const cpuModuleId = configElement.getAttribute('ModuleId') ?? undefined;
+    const cpuModuleId = configElement?._att?.ModuleId as unknown;
     // Get properties from <AutomationRuntime> element
-    const arElement = getChildElements(configElement, 'AutomationRuntime').pop();
-    const arVersion = arElement?.getAttribute('Version') ?? undefined;
+    const arVersion = configElement?.AutomationRuntime?._att?.Version as unknown;
     // Get properties from <Build> element
-    const buildElement = getChildElements(configElement, 'Build').pop();
-    const gccVersion = buildElement?.getAttribute('GccVersion') ?? undefined;
-    const buildOptionsRaw = buildElement?.getAttribute('AdditionalBuildOptions');
-    const buildOptions = splitShellArgs(buildOptionsRaw);
-    const ansiCBuildOptionsRaw = buildElement?.getAttribute('AnsicAdditionalBuildOptions');
-    const ansiCBuildOptions = splitShellArgs(ansiCBuildOptionsRaw);
-    const iecBuildOptionsRaw = buildElement?.getAttribute('IecAdditionalBuildOptions');
-    const iecBuildOptions = splitShellArgs(iecBuildOptionsRaw);
-    const ansiCIncludeDirsRaw = buildElement?.getAttribute('AnsicIncludeDirectories');
-    const ansiCIncludeDirs = parseIncludeDirs(ansiCIncludeDirsRaw);
+    const gccVersion = configElement?.Build?._att?.GccVersion as unknown;
+    const buildOptionsRaw = configElement?.Build?._att?.AdditionalBuildOptions as unknown;
+    const buildOptions = typeof buildOptionsRaw === 'string' ? splitShellArgs(buildOptionsRaw) : [];
+    const ansiCBuildOptionsRaw = configElement?.Build?._att?.AnsicAdditionalBuildOptions as unknown;
+    const ansiCBuildOptions = typeof ansiCBuildOptionsRaw === 'string' ? splitShellArgs(ansiCBuildOptionsRaw) : [];
+    const iecBuildOptionsRaw = configElement?.Build?._att?.IecAdditionalBuildOptions as unknown;
+    const iecBuildOptions = typeof iecBuildOptionsRaw === 'string' ? splitShellArgs(iecBuildOptionsRaw) : [];
+    const ansiCIncludeDirsRaw = configElement?.Build?._att?.AnsicIncludeDirectories as unknown;
+    const ansiCIncludeDirs = typeof ansiCIncludeDirsRaw === 'string' ? splitShellArgs(ansiCIncludeDirsRaw) : [];
     // function to resolve include URIs from project root
     const resolveAnsiCIncludeDirs = (projectRoot: Uri) => {
         return ansiCIncludeDirs.map((path) => pathResolve(projectRoot, path));
     };
     // return info data
     return {
-        arVersion: arVersion,
-        cpuModuleId: cpuModuleId,
+        arVersion: typeof arVersion === 'string' ? arVersion : undefined,
+        cpuModuleId: typeof cpuModuleId === 'string' ? cpuModuleId : undefined,
         build: {
-            gccVersion: gccVersion,
+            gccVersion: typeof gccVersion === 'string' ? gccVersion : undefined,
             additionalBuildOptions: buildOptions,
             ansiCAdditionalBuildOptions: ansiCBuildOptions,
             iecAdditionalBuildOptions: iecBuildOptions,