+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Add new but unreleased features, fixes... here during development
 
 Users can look forward to these changes for the next release 😀😉
 
+### Changed
+
+- Changed internal XML backend from `xmlbuilder2` to `fast-xml-parser`. This change will allow us to easier parse and also modify XML files for new features.
+- Improved AS V3.x configuration view package files support. We still (and maybe will never) not support all settings and files.
+
 ### Fixed
 - The default connection port in the ArSim transfer task was adjusted.
   [#39](https://github.com/br-automation-com/vscode-brautomationtools/issues/39)<br/>